Changed config values from float to int

+++ common/src/main/java/net/moritz_htk/music_delay_reducer/config/MDRConfig.java
@@ -1,6 +1,6 @@
 package net.moritz_htk.music_delay_reducer.config;
 
 public interface MDRConfig {
-    float minTime();
-    float maxTime();
+    int minTime();
+    int maxTime();
 }
